requirements.txt
absl-py
astunparse
blessings
cachetools
certifi
chardet
cycler
docopt
flatbuffers
gast
google-auth
google-auth-oauthlib
google-pasta
gpustat
grpcio
grpcio-tools
h5py
idna
importlib-metadata
Keras-Preprocessing
kiwisolver
Markdown
matplotlib
numpy
nvidia-ml-py3
oauthlib
opencv-python
opencv-python-headless
opt-einsum
pandas
Pillow
protobuf
psutil
pyasn1
pyasn1-modules
pyparsing
python-dateutil
pytz
requests
requests-oauthlib
rsa
scipy
six
tensorboard
tensorboard-plugin-wit
tensorflow
tensorflow-estimator
tensorflow-serving-api
termcolor
typing-extensions
urllib3
Werkzeug
wrapt
yarg
zipp
